Add --skip-npm flag
Have Lerna update all versions and dependencies where required in
`package.json` files without actually publishing the changes to NPM.

This is useful for publishing manually, which is necessary to do periodically
to get README updates to show up on npmjs.com
(until https://github.com/npm/newww/issues/389 is resolved).

This is a workaround for #64.

This was proposed in #127.