Diff compaction for package binary PR
This is an attempt to get @rygine's #192 merged by trimming extraneous changes
to slim down the diff and ease review.

None of the following changes reflect _disapproval_ of the code in the
original PR.  This is pure change-set minimization.

Changes:
- Revert incidental code reorganization in:
    - `createLinkDependency`
    - `linkDependencies`
    - `createLinkedDependencyFiles`
- Thin interface of `createBinaryLink`
- Eliminate unused optional argument to `FileSystemUtilities.symlink`
- Simplify `createBinaryLink`
- Remove jsdoc comments

With this patch the diff against master is now pure addition.