Handle dependency graph cycles during bootstrap
The dependency-order bootstrap scheduler introduced in #250 assumes that the
dependency graph is acyclic.  This turns out not to be a valid assumption, as
demonstrated in #299.

This patch falls back to the package with the most dependents when bootstrap
is blocked by a cycle in the dependency graph.  It issues a warning that graph
cycles may cause instability during bootstrap if dependencies are used during
`prepublish`.