Make log levels more like npm
Currently Lerna has log _types_ and a logging verbose flag, which has
different default values for different types (e.g. `false` for "debug",
`true` for "error").  It also has log types that are close to, but don't
_quite_ match their npm counterparts.

This PR makes Lerna's logging configuration more similar to npm's.

It:

- Adds a `--loglevel` CLI option
- Drops the `verbose` argument to logger methods
- Renames some logger methods to be more like their `npm` equivalents
    - `logger.debug` => `logger.verbose`
    - `logger.warning` => `logger.warn` (also more like `console.warn`)
- Allows an optional error object to be included with any log level

The update to the README is basically cribbed from `npm help 7 config`.

Differences with `npm`:
- Lerna has a `success` log level in place of npm's `http`
- Lerna's default is `info`, while npm's default is `warn`