Automatic hoisting of common dependencies to repo root (#36)
* Resurrect @rygine's groundwork for common dep hoisting

This is taken from the base commit on the "enhanced bootstrap" PR.

* Whitespace for legibility

Easier on my old eyes while I'm trying to grok this. :)

* Update a logger statement

This method was renamed while this change set was in flight.

* Always warn on un-shared dep (even if already installed)

* Only install deps in the root if not already satisfied

* Simplify most common version determination

* Simplify getDependenciesToInstall

Mostly just so I'm sure I understand it.

* Pass dependents back from getDependenciesToInstall

We'll need those to link binaries from the root.

* Actually hoist

* Update commentary

* Link binaries for hosted dependencies

* Don't override an explicit root dependency

* Gate hoisting with --hoist and --nohoist options

* Update a test to expect slight change to install order

* Add a test and fix a bug it uncovered :D

Scoped packages weren't hoisted with `--hoist`.

* Add a test for --nohoist

* Prune hoisted deps from package dirs

* Factor out dependent package fetch

* Add some documentation about hoisting

* Root version may have no dependents

If it's depended on by the root package.json and others don't match.

* Only warn about mismatch with root if ther's a root version

* Start root installs first, since they're likely to take longest

* Make tests expect root install first

* Don't include null deps in root install

These will be present when deps are already satisfied at the root.

* Kill a vestigial comment