fix: always install leaves using `npm --global-style`
In the event that an explicit dependency (`dep1@X`) of a package is
hoisted but other dependencies with a common sub-dependency on a
different version of `dep1@Y` cannot be hoisted, npm will attempt to
share that package by installing `dep1@Y` in the root of the
package's `node_modules` directory. As a result, the incorrect
version of dep1 will be available to the package rather than the
hoisted dependency it requires.

To avoid this, always use `npm install --global-style` for "leaves"
in a hoisted context. This results in more consistently correct
behavior at the cost of increased duplication of some packages.

See: https://github.com/lerna/lerna/issues/677