Add --parallel flag to `lerna run`
This allows simpler invocation of `watch` scripts, with the caveat that
concurrency and topological sorting are _completely_ ignored.
This is generally the intention when calling `lerna run watch` and other
similar script targets, hence the additional flag.

    # the following commands are equivalent
    lerna run watch --concurrency=1000 --stream
    lerna run watch --parallel

Package filtering (`--scope` and `--ignore`) is still available when this
new flag is being used, and it is advised to narrow the scope of parallel
execution when you have more than a dozen packages or so (YMMV).