updated: return non-zero exit code when no changes
Instead of throwing an error we really just want to return a non-zero
exit code.

This patch adds the possibility to return a custom exit code (with a
default of zero) and utilizes that possibility to return 1 as a exit
code when a user runs the `updated` command and there are no packages
that need updating.

Related-to: #802