publish: return zero exit code instead of error
Instead of throwing a error we want to just output a message that there
is nothing for `publish` to do at this time since there are no changes.
This helps in CI systems when you'd just want to run `lerna publish`
without having to conditionally running the command based on if there
are any changes.

This changes the exit code so it would be a breaking change for anyone
that currently relies on the exit code to be non-zero.

Related-to: #802