fix: Fix issue where Yargs default would override durable options.
The solution to this requires that we stop using yargs defaults because we have several different ways of specifying options. If we specify yargs defaults, then they get set as flags and override any specified in the lerna.json. To get around this, we have to put defaults last in the chain by specifying them per command, as opposed to yargs which comes first in the chain as they merge user-defined options passed as flags with the corresponding defaults. There's no way for us to insert our lerna.json options into this at runtime.