Add a link command to the CLI - resolves #341
This commit adds a new 'link' command to the CLI which will create symlinks across the lerna package dependencies. This command was requested in #341 and solves issues with using 'npm prune' with lerna based repositories, e.g. 'npm prune' removes extranous packages, which include the symlinks of lerna.

Since the bootstrap command is a bit heavy, especially while hoisting, I opted for a seperate command since the bootsrap command is pretty complex as of now anyway and I don't think it will be particularly usefull to add option after option.

I also refactored most of the code, e.g. made it more purely functional, regarding the symlinkPackages method and extracted it into the PackageUtilities