Truncate stack (or pass directly) when logging errors
npmlog has a weird habit of repeating itself when serializing error objects
passed to any level method (e.g., `log.error()`). To avoid the stutter, we
must pass the stack directly (if it exists).

To address needless verbosity in the now-deduplicated stacks, do a little
string splitting to only output the stack trace from the beginning of the
command execution lifecycle. This keeps yargs out of the stack, for example.