feat: Generate single root changelog in fixed mode
When using lerna in fixed mode it would be desireable to have lerna
generate a consolidated changelog.md file in the root directory that
contains changes for all packages.

This commit adds said feature, so that, when executing
`lerna publish --conventional-commits` it will generate one changelog
per package and also one consolidated changelog in the root of the
repository.
The changelog generation in independent mode stays the same (it
generates one changelog per package).

Fixes: #1015