Use local lerna if available (#1122)
Closes #138

BREAKING CHANGE: lerna now prefers installs resolvable relative to the current working directory.

For situations where a global install was used and an outdated lerna version is present in the node_modules folder, this might cause breakage. The issue can be resolved by installing a lerna version that fulfills the version requirement in lerna.json.