Add post init documentation for independent mode
If merged this will add a small bit of documentation telling you how to make your existing lerna project use independent versioning.

This was something I had to do in our application, and I just figured it out by doing a `lerna init --independent` on a test folder, and seeing what the difference was with my current `lerna.json` file. 

Fortunately, it worked. But I feel it'd be for the best if this were documented, and we could save some people some searching in the future.