PackageGraph is a Map, its nodes store relationship Sets (#1260)
A lot of existing code was re-calculating stuff that should live inside the PackageGraph; namely, the relationships between packages (localDependencies and localDependents). I made PackageGraph a subclass of Map, mostly because it already had a `.get()` method, and now it can be iterated and such. It's a lot more elegant, overall.

A big piece of the new relationship logic is based on [`npm-package-arg`](https://github.com/npm/npm-package-arg#result-object) parsing, which resolves a lot of previously half-baked logic. It's no longer necessary to pass the git version parser to the graph constructor, for example. `npm-package-arg` will also facilitate future graph manipulations, such as recognizing relative `file:` specifiers correctly.

### `Package`

* The package `bin` field now properly resolves an object when the original value is a string shorthand.
* Many `Package` properties are immutable now, and very few are actually enumerable.

### `PackageGraph`

* `--force-local` evaluation has been moved into `PackageGraph`'s constructor, simplifying the existing evaluation of local dependencies during symlink creation in `LinkCommand`.
* Cycle detection and removal has been moved into a `PackageGraph` instance method, making the topological batching algogrithm much easier to read.
* `PackageGraphNode`s no longer have a confusingly-named "dependencies" list, but distinct `Map` instances `.localDependencies` and `.localDependents`.
* `PackageGraphNode`s themselves now forward most of the interesting `Package` properties, so the potential is there to use them directly, instead of the myriad places we pluck a (renamed) `pkg` property from various containers.

### `UpdatedPackagesCollector`

* Now reuses all of the new graph calculation in `PackageGraph` instead of blithely duplicating it.
* There's a lot of room for other improvements in this utility.

### `PackageUtilities`

* Renamed `topologicallyBatchPackages()` to `batchPackages()`, because ain't nobody got time for that many syllables.