feat(amend): new --amend flag for the publish command
For the CI/CD use case of releasing-per-commit, adding a commit to update
the version number and/or the CHANGELOG, for every merged patch, creates
unnecessary noise. This patch allows us to amend the existing commit with any
new changes, without replacing the message or generating a new commit.

Specific changes:
- Added the --amend commandline flag to the publish command
- Add an 'amend' flag to the git-commit helper, which translates into `--amend --no-edit`.
- Added tests for all the above.
- Updated the README.md flag with the new option.
- Two Appveyor tweaks for environment parameters, one for the locale, the other for node feature flags.