feat: add git-verify flag to bypass --no-verify
Some setups (e.g. Gerrit) require commit hooks that add a change id.

So that lerna publish can properly push the commit in these
environments, the `--no-verify` commit flag should not be specified.
This adds a new flag `git-verify`, that if set to `true` will run the
commit command without that option, so that all hooks are run.