fix(publish): move git sha to build metadata
Move the git commit sha from the pre-release string of the semantic
version number to the build metadata string.

Git commit shas are not always valid pre-release strings according to
the Semantic Versioning Specification 2.0. For example, SemVer requires
that a numeric pre-release segment (a segment consisting of only
numbers) must not contain a leading zero. However, it's possible for
a git commit sha to be all numbers and for the first number to be zero.

The same rule does **not** apply to the build metadata string placed
after a `+` character. Therefore, to ensure the version contains git sha
information, but is still valid, we are moving the sha to the build
metadata string.

BREAKING CHANGE:

Git commit sha is now contained in the build metadata string of the
semantic version number.

So `1.1.0-alpha.81e3b443` now becomes `1.1.0-alpha+81e3b443` when you
run `lerna --canary`.

To increment the pre-release number after `alpha` please include it in
the string passed to the `--canary` flag:

```bash
lerna --canary=alpha.0
```

For convience, consider using the build number typically set in your
build environment by your CI platform:

```bash
lerna --canary=alpha.${BUILD_NUMBER}
```

Fixes #1118