fix: properly update lockfile v2
npm's lockfile v2 adds a new top-level key: packages. This is not really
noteworthy except for the fact that the first entry in there is a copy
of the package this lockfile is for, stored in the key `""`, and this
copy _also_ has a version key. This commit checks for and updates that
key (`packages."".version`) if it exists.

Fixes #2832.