feat(version): add detect-preid option
The current behaviour when incrementing prerelease does not take into
account existing versions. This can lead to git tag failing when the
resulting prerelease version already exists.

For example, given an alpha branch with version 1.0.1-alpha.0 is merged
to a beta branch with version 1.0.1-beta.4 (previous versions included).
The next version of the beta branch becomes 1.0.1-beta.0, an already
existing version and tag.

A version option --detect-preid will attempt to find and bump from a
previous existing prerelease version, if it exists. This is limited to
version bumps using conventional commits.