fix file formatting on git add
PR #3766 brought prettier 3 support by using asynchronous prettier APIs,
but missed one `await` when calling `.format(...)`.

Because of this, `fs.writeFileSync(...)` is called with a Promise
instead of a string and throws an exception:

> TypeError [ERR_INVALID_ARG_TYPE]: The "data" argument must be of type
> string or an instance of Buffer, > TypedArray, or DataView. Received
> an instance of Promise

This exception is caught by the try/catch, and the file is never
written.

Worse, when formatting files that are not supported by prettier (ex:
yarn.lock), `.format(...)` fails and the promise rejection isn't
handled, which lead to the following exception:

> UndefinedParserError: No parser could be inferred for file "..."

and Node terminates the process, which is probably the cause of #3788